Дорабоки для локального бота

+++ bitbucketbot-core/src/main/java/org/sadtech/bot/vcs/bitbucket/core/AbstractPullRequestBitbucketParser.java
@@ -55,7 +55,7 @@ public abstract class AbstractPullRequestBitbucketParser implements PullRequestP
         log.info("Открыты: " + Arrays.toString(openId.toArray()));
         log.info("Закрыты: " + Arrays.toString(closeId.toArray()));
         log.info("Не найдены: " + Arrays.toString(newNotExistsId.toArray()));
-        if (!newNotExistsId.isEmpty()) {
+        if (!newNotExistsId.isEmpty() && !openId.isEmpty()) {
             pullRequestsService.deleteAllById(newNotExistsId);
         }
     }
